wly: Set stock brightness levels
Minimum brightness is obtained from stock by disabling auto-brightness
and operating the brightness slider to minimum with running
cat /sys/class/backlight/panel0-backlight/brightness

This fetches a value of 4, the value that we currently have
is 4/2047. However, the display dtsi exposes the range of
1 - 4095, so 4/4095 is being applied. The same value is
applied for both dim and doze.

Default brightness setting is applied as 1023/4095

Change-Id: I0b2513be866728f3ddf93f7fbe296d78d862c813